reduce imports, revert accordion changes, rename components
